requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
Werkzeug==3.1.4
